Give mail_banned a default value in the account database
The mail_banned field is set as "NOT NULL", but is not given a value by
SmrAccount::createAccount. This causes account creation to fail the
INSERT database query.

The default value is set as zero. Since mail_banned is compared to TIME,
in SmrAccount::isMailBanned, this value corresponds to no mail ban.