Give message_notifications a default value in the account database
The message_notifications field is set as "NOT NULL", but is not given
a value by SmrAccount::createAccount. This causes account creation to
fail the INSERT database query.

We allow the value to be NULL, which should set it as an empty object
when SmrAccount is constructed. The getter/setter methods check to
see if the keys of this object are set, and do the appropriate thing
if they are not.