Give hotkeys a default value in the account database
The hotkeys field is set as "NOT NULL", but is not given a value by
SmrAccount::createAccount. This causes account creation to fail the
INSERT database query.

We allow the value to be NULL, which should set it as an empty object
when first queried by getObject in SmrAccount::__construct, but then
it is immediately set to the default hotkeys on the following line.