shop_weapon.php: use list.js to sort weapons for sale
This is useful primarily when you have a large list of weapons
(e.g. Race Wars Weapons) and you want to figure out which weapon
to buy.

It is not ideal for buying the same weapon multiple times,
because after you buy a weapon the sorting reverts.
I don't think it's a blocking issue due to the limited use cases.

(Perhaps in the future, more globally, we can have some system
in place to preserve client-side sorting using ajax and sessions,
for example.)