Modify port raid alignment change
Players wanted a more expeditious way to change alignment,
since player kills are unreliable and port raids are
excruciatingly slow (`2*Port Level` for only one raider
for each successful bust).

This patch implements the following 3 changes:

* We need the alignment change to occur for all participating
  players, otherwise there is too much randomness.

* We need the change to occur on every shot so that there is
  some relationship between the number of turns used and
  the size of the alignment change (e.g. we don't want people
  to game the system by coming in for the kill shot).

* We implement a damage threshold for the alignment change
  for two reasons.
    1) With a +/- 1 change on every shot for every player,
       alignment would change too rapidly. So without converting
       alignment to a float and using fractional increments,
       we need some other way to scale everything down.
    2) Without a damage threshold, players could spam a Level 1
       port with a single small weapon to artificially increase
       the number of shots per port to an absurd number. We
       want to ensure the PR's are at least somewhat legitimate.

The threshold is currently set at 300 damage, so that if your
Nuke/HHG hits, you automatically get the alignment change.
This is higher than the typical average damage per shot, but
that is needed to avoid turbo gains. This number should be
tweaked according to player feedback.