Stay on Local Map when dropping single forces
Previously, dropping a single force using the "[x]" links on the
right panel would always take you to the Current Sector page.

However, if you are on the Local Map, it is likely that you want
to stay on the Local Map, and so we now do that in order to avoid
unnecessary clicking and page switching.

This may have some impact on gameplay, since the old behavior
does combine two actions together in a relevant way (e.g. dropping
a single mine and then going to the Current Sector page to attack).
We will want to monitor the community feedback from this change
before deploying it to the live server.